[Fe][Feat] #145 : 피드백 사항 수정
